Fix for Mac OS X compatibility problem.
SF_INFO::frames is of type sf_count_t, which is a 32bit int on most
platforms, but 64bit on Mac OS X. To pass it as S_INTEGER_PARAM_TYPE,
I have to make an explicit cast to int. Note that this is not the
correct thing to do, though. I left a FIXME note to that respect.